IPRT: More VC++ 14.1 build fixes.  bugref:8489
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@83739 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Runtime/common/fs/isovfs.cpp
@@ -833,9 +833,9 @@ static int rtFsIsoCore_InitFrom9660DirRec(PRTFSISOCORE pCore, PCISO9660DIRREC pD
             }
             else
             {
-                size_t cbSkip = (offDirRec + pVol->cbSector) & ~(pVol->cbSector - 1U);
+                uint64_t cbSkip = (offDirRec + pVol->cbSector) & ~(uint64_t)(pVol->cbSector - 1U);
                 offDirRec += cbSkip;
-                pDirRec = (PCISO9660DIRREC)((uintptr_t)pDirRec + cbSkip);
+                pDirRec = (PCISO9660DIRREC)((uintptr_t)pDirRec + (size_t)cbSkip);
             }
         }
     }
